Publish to analytics + simple status code
Added support for using simple return codes i.e. only the fact if a
response code was successful or not is evaluated. Hence the return of
Success will imply the return code "1" - and everything else (UNKNOWN,
CANCELED, FAILED, ERROR) will return 0.

This makes it easier to make calculations based on the Status field.

Added support for pushing data to Analytics - in addition to still
pushing data to the metric store.
Pushing URL monitor data to analytics enables XLM for the data points.

Added a new section (analyticsConfig) to the config.xml that controls
the connection to analytics
	useSimpleReturnCodes: true
    publishToAnalytics:   true
    analyticsApiUrl:      https://analytics.api.appdynamics.com
    globalAccountName:    <global account name>
    apiKey:               <analytics api key>
    analyticsSchemaName:  <name of custom schema - for example:
UrlMonitorSchema>